Moves Annuity/StraightLine tests from 'se' -> 'common'
