+++ src/test/kotlin/org/openbroker/common/model/AnnuityTotalCostForLoanTest.kt
@@ -1,61 +1,60 @@
-package org.openbroker.se.privateunsecuredloan.model
+package org.openbroker.common.model
 
 import org.junit.jupiter.api.Assertions.assertEquals
 import org.junit.jupiter.api.Test
 import org.openbroker.common.loancosts.Annuity
-import org.openbroker.common.loancosts.StraightLine
 import java.math.BigDecimal
 
-private const val ACCEPTABLE_ERROR: Double = 0.5
+private const val ACCEPTABLE_ERROR: Double = 0.2
 
-class StraightLineTotalCostForLoanTest {
+class AnnuityTotalCostForLoanTest {
 
     @Test
     fun testComputeTotalCostForSmallLoanNoTermFee() {
-        val totalCost: BigDecimal = StraightLine.totalCost(
+        val totalCost: BigDecimal = Annuity.totalCost(
             loanAmount = 10_000,
             nominalAnnualInterestRate = 0.05,
             termFee = 0,
             termMonths = 24
         )
 
-        assertEquals(10_521.0, totalCost.toDouble(), ACCEPTABLE_ERROR)
+        assertEquals(10_529.0, totalCost.toDouble(), ACCEPTABLE_ERROR)
     }
 
     @Test
     fun testComputeTotalCostForBigLoanNoTermFee() {
-        val totalCost: BigDecimal = StraightLine.totalCost(
+        val totalCost: BigDecimal = Annuity.totalCost(
             loanAmount = 200_000,
             nominalAnnualInterestRate = 0.067,
             termFee = 0,
             termMonths = 96
         )
 
-        assertEquals(254_158.0, totalCost.toDouble(), ACCEPTABLE_ERROR)
+        assertEquals(258_910.0, totalCost.toDouble(), ACCEPTABLE_ERROR)
     }
 
 
     @Test
     fun testComputeTotalCostForSmallLoanWithTermFee() {
-        val totalCost: BigDecimal = StraightLine.totalCost(
+        val totalCost: BigDecimal = Annuity.totalCost(
             loanAmount = 10_000,
             nominalAnnualInterestRate = 0.05,
             termFee = 20,
             termMonths = 24
         )
 
-        assertEquals(11_001.0, totalCost.toDouble(), ACCEPTABLE_ERROR)
+        assertEquals(11_009.0, totalCost.toDouble(), ACCEPTABLE_ERROR)
     }
 
     @Test
     fun testComputeTotalCostForBigLoanWithTermFee() {
-        val totalCost: BigDecimal = StraightLine.totalCost(
+        val totalCost: BigDecimal = Annuity.totalCost(
             loanAmount = 200_000,
             nominalAnnualInterestRate = 0.067,
             termFee = 60,
             termMonths = 96
         )
 
-        assertEquals(259_918.0, totalCost.toDouble(), ACCEPTABLE_ERROR)
+        assertEquals(264_670.0, totalCost.toDouble(), ACCEPTABLE_ERROR)
     }
 }
